Move SwapAmounts to own module
