+++ swap/src/alice/swap.rs
@@ -16,7 +16,7 @@ use crate::{
     state,
     state::{Alice, Swap},
     storage::Database,
-    SwapAmounts,
+    swap_amounts::SwapAmounts,
 };
 use anyhow::{bail, Result};
 use async_recursion::async_recursion;
