+++ swap/src/lib.rs
@@ -1,7 +1,6 @@
 #![allow(non_snake_case)]
 
-use ::serde::{Deserialize, Serialize};
-use std::fmt::{self, Display};
+use crate::swap_amounts::SwapAmounts;
 
 pub mod alice;
 pub mod bitcoin;
@@ -12,6 +11,7 @@ pub mod network;
 pub mod serde;
 pub mod state;
 pub mod storage;
+pub mod swap_amounts;
 pub mod trace;
 
 #[cfg(test)]
@@ -30,28 +30,4 @@ pub enum Cmd {
 pub enum Rsp {
     VerifiedAmounts,
     Abort,
-}
-
-/// XMR/BTC swap amounts.
-#[derive(Copy, Clone, Debug, Serialize, Deserialize, PartialEq)]
-// TODO(Franck): review necessity of this struct
-pub struct SwapAmounts {
-    /// Amount of BTC to swap.
-    #[serde(with = "::bitcoin::util::amount::serde::as_sat")]
-    pub btc: bitcoin::Amount,
-    /// Amount of XMR to swap.
-    #[serde(with = "xmr_btc::serde::monero_amount")]
-    pub xmr: monero::Amount,
-}
-
-// TODO: Display in XMR and BTC (not picos and sats).
-impl Display for SwapAmounts {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(
-            f,
-            "{} sats for {} piconeros",
-            self.btc.as_sat(),
-            self.xmr.as_piconero()
-        )
-    }
 }
