+++ swap/src/network/request_response.rs
@@ -1,3 +1,4 @@
+use crate::swap_amounts::SwapAmounts;
 use async_trait::async_trait;
 use futures::prelude::*;
 use libp2p::{
@@ -7,8 +8,6 @@ use libp2p::{
 use serde::{Deserialize, Serialize};
 use std::{fmt::Debug, io, marker::PhantomData};
 use tracing::debug;
-
-use crate::SwapAmounts;
 use xmr_btc::{alice, bob, monero};
 
 /// Time to wait for a response back once we send a request.
