+++ swap/src/swap_amounts.rs
@@ -1,0 +1,27 @@
+use crate::monero;
+use ::serde::{Deserialize, Serialize};
+use std::fmt::{self, Display};
+
+/// XMR/BTC swap amounts.
+#[derive(Copy, Clone, Debug, Serialize, Deserialize, PartialEq)]
+// TODO(Franck): review necessity of this struct
+pub struct SwapAmounts {
+    /// Amount of BTC to swap.
+    #[serde(with = "::bitcoin::util::amount::serde::as_sat")]
+    pub btc: bitcoin::Amount,
+    /// Amount of XMR to swap.
+    #[serde(with = "xmr_btc::serde::monero_amount")]
+    pub xmr: monero::Amount,
+}
+
+// TODO: Display in XMR and BTC (not picos and sats).
+impl Display for SwapAmounts {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(
+            f,
+            "{} sats for {} piconeros",
+            self.btc.as_sat(),
+            self.xmr.as_piconero()
+        )
+    }
+}
